refactor(113181): refatora lugar do componente
