+++ src/components/screens/Cadastros/EditaisContratosRefatorado/Cadastro/index.tsx
@@ -27,7 +27,7 @@ import { getLotesSimples } from "services/lote.service";
 import { getNomesTerceirizadas } from "services/produto.service.js";
 import { FormCadastroEditaisContratosInterface } from "../interfaces";
 import "./style.scss";
-import { FieldArrayContratos } from "./components";
+import { FieldArrayContratos } from "./components/FieldArrayContratos";
 import { ModalCadastroEdital } from "./ModalCadastroEdital";
 import {
   CADASTROS,
